- Added gtk libraries into package - moved cli functions to cli.c, gcs_gui can now be used on the cli
