+++ .github/workflows/build-test.yml
@@ -13,7 +13,7 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: [ubuntu-latest, macos-latest, windows-latest]
+        os: [ubuntu-latest, windows-latest]
 
     steps:
     - uses: actions/checkout@v4
@@ -24,11 +24,6 @@ jobs:
       run: |
         sudo apt-get update
         sudo apt-get install -y libgtk-3-dev pkg-config
-
-    # --- GTK3 Dependencies (macOS) ---
-    - name: Install GTK3 (macOS)
-      if: runner.os == 'macOS'
-      run: brew install gtk+3 pkg-config
 
     # --- Windows Setup ---
     # Installs MSYS2 to provide GCC, POSIX tools (sh, make), and GTK3 on Windows
