+++ .github/workflows/release.yml
@@ -19,17 +19,12 @@ jobs:
             cli_artifact: gcs
             cli_asset: gcs-linux-x64
             gui_artifact: gcs-gui
-            gui_asset: gcs-gui-linux-x64
-          - os: macos-latest
-            cli_artifact: gcs
-            cli_asset: gcs-macos-arm64
-            gui_artifact: gcs-gui
-            gui_asset: gcs-gui-macos-arm64
+            gui_zip: gcs-gui-linux-x64.zip
           - os: windows-latest
             cli_artifact: gcs.exe
             cli_asset: gcs-windows-x64.exe
             gui_artifact: gcs-gui.exe
-            gui_asset: gcs-gui-windows-x64.exe
+            gui_zip: gcs-gui-windows-x64.zip
 
     steps:
       - uses: actions/checkout@v4
@@ -40,11 +35,6 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y libgtk-3-dev pkg-config
-
-      # --- GTK3 Dependencies (macOS) ---
-      - name: Install GTK3 (macOS)
-        if: runner.os == 'macOS'
-        run: brew install gtk+3 pkg-config
 
       # --- Windows Setup ---
       - name: Setup MSYS2 (Windows)
@@ -65,13 +55,41 @@ jobs:
         shell: msys2 {0}
         run: make
 
-      # --- Packaging ---
-      # Rename the raw binaries to the final asset names
-      - name: Rename Artifacts
+      # --- Packaging Pure CLI ---
+      - name: Rename CLI Artifact
+        shell: bash
+        run: mv ${{ matrix.cli_artifact }} ${{ matrix.cli_asset }}
+
+      # --- Packaging GUI & Libs (Linux) ---
+      - name: Bundle GUI Dependencies (Linux)
+        if: runner.os == 'Linux'
+        run: |
+          mkdir -p dist/libs
+          mv ${{ matrix.gui_artifact }} dist/gcs
+          # Grab dynamically linked libraries
+          ldd dist/gcs | grep "=> /" | awk '{print $3}' | xargs -I '{}' cp -n '{}' dist/libs/ || true
+          # Create a launcher script to tell Linux to use the local libs
+          echo '#!/bin/bash' > dist/run_gcs.sh
+          echo 'DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"' >> dist/run_gcs.sh
+          echo 'LD_LIBRARY_PATH="$DIR/libs" "$DIR/gcs" "$@"' >> dist/run_gcs.sh
+          chmod +x dist/run_gcs.sh
+
+      # --- Packaging GUI & Libs (Windows) ---
+      - name: Bundle GUI Dependencies (Windows)
+        if: runner.os == 'Windows'
+        shell: msys2 {0}
+        run: |
+          mkdir -p dist/libs
+          mv ${{ matrix.gui_artifact }} dist/gcs.exe
+          # Grab all MSYS2 MinGW DLLs linked to the executable
+          ldd dist/gcs.exe | grep '\/mingw.*\.dll' -o | xargs -I{} cp {} dist/libs/ || true
+
+      # --- Create Zip Archive ---
+      - name: Zip GUI Bundle
         shell: bash
         run: |
-          mv ${{ matrix.cli_artifact }} ${{ matrix.cli_asset }}
-          mv ${{ matrix.gui_artifact }} ${{ matrix.gui_asset }}
+          cd dist
+          zip -r ../${{ matrix.gui_zip }} ./*
 
       # --- Uploading ---
       - name: Create Release & Upload
@@ -80,4 +98,4 @@ jobs:
         with:
           files: |
             ${{ matrix.cli_asset }}
-            ${{ matrix.gui_asset }}
+            ${{ matrix.gui_zip }}
